fall back to attribute selector for getElementById
This adds a function that detects whether a given document (or fragment)
has the `getElementById()` method and catches errors if the function
fails, then falls back on a call to `querySelector()` with an ID
selector in the form `[id="{id}"]` (with quotes escaped).

This function is called for both the global `document` and for document
fragments loaded via XHR, since document fragments lack
`getElementById()` support in many browsers (namely, IE and Edge), and
valid IDs may contain characters that incorrectly trigger false
negatives in even modern implementations of `getElementById()`.

Fixes #73.